requirements.txt
opencv-python==4.5.5.64
matplotlib==3.6.0
numpy==1.23.3
easyocr==1.6.2
paho-mqtt==1.6.1
Pillow==9.3.0
